students/K33392/Taiakin_Daniil/lr2/task2/async_conn.py
import os
from dotenv import load_dotenv
from sqlalchemy.ext.asyncio import create_async_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from sqlalchemy.ext.asyncio import AsyncSession

load_dotenv()

engine = create_async_engine(os.environ["SQLALCHEMY_DATABASE_URL"])

asyncSessionLocal = sessionmaker(
    class_=AsyncSession, autocommit=False, autoflush=False, bind=engine
)

Base = declarative_base()

async def init_db():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.create_all)


async def get_session():
    async with asyncSessionLocal() as session:
        yield session